refactor(lua): replace vscode-neovim with vscode
